extract file type detection into decoupledLogic
