+++ web/src/core/usecases/dataExplorer/decoupledLogic/index.ts
@@ -1,0 +1,6 @@
+export {
+    detectFileTypeFromFileDownload,
+    detectFileTypeFromSourceUrlExtension
+} from "./detectFileType";
+
+export { isValidFileType, type ValidFileType } from "./fileType";
